Implementar método deletar guilda para validar se ela existe

+++ src/main/java/com/br/zup/gerenciadordeguildas/services/GuildaService.java
@@ -1,6 +1,7 @@
 package com.br.zup.gerenciadordeguildas.services;
 
 import com.br.zup.gerenciadordeguildas.entities.Guilda;
+import com.br.zup.gerenciadordeguildas.exceptions.RecursoNaoEncontradoException;
 import com.br.zup.gerenciadordeguildas.repositories.GuildaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,6 +21,10 @@ public class GuildaService {
     }
 
     public void deletarGuildas(Integer id) {
-        guildaRepository.deleteById(id);
+        if(guildaRepository.existsById(id)){
+            guildaRepository.deleteById(id);
+        }
+
+        throw new RecursoNaoEncontradoException("Guilda", id);
     }
 }
